# swagger-openapi.json
{"openapi":"3.1.0","info":{"title":"AutoCheckMobile Backend","description":"REST API AutoCheckMobile. Все ответы отдаются в формате `{data, error, meta}`. Все эндпоинты требуют Bearer Token, кроме `/auth/login` и `/auth/register`.","contact":{"name":"AutoCheckMobile Team"},"version":"1.0.0"},"paths":{"/api/v1/auth/register":{"post":{"tags":["auth"],"summary":"Регистрация пользователя","operationId":"register_api_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth"],"summary":"Авторизация пользователя","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["auth"],"summary":"Выход пользователя","operationId":"logout_api_v1_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/auth/profile":{"get":{"tags":["auth"],"summary":"Профиль текущего пользователя","operationId":"profile_api_v1_auth_profile_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/assignments":{"get":{"tags":["assignments"],"summary":"Список тестовых заданий","operationId":"list_assignments_api_v1_assignments_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["assignments"],"summary":"Создание тестового задания","operationId":"create_assignment_api_v1_assignments_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignmentCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/assignments/{assignment_id}":{"get":{"tags":["assignments"],"summary":"Получить задание по ID","operationId":"get_assignment_api_v1_assignments__assignment_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"assignment_id","in":"path","required":true,"schema":{"type":"integer","title":"Assignment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["assignments"],"summary":"Редактировать задание","operationId":"update_assignment_api_v1_assignments__assignment_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"assignment_id","in":"path","required":true,"schema":{"type":"integer","title":"Assignment Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignmentUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["assignments"],"summary":"Удалить задание","operationId":"delete_assignment_api_v1_assignments__assignment_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"assignment_id","in":"path","required":true,"schema":{"type":"integer","title":"Assignment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/submissions":{"get":{"tags":["submissions"],"summary":"Список проверок","operationId":"list_submissions_api_v1_submissions_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["submissions"],"summary":"Загрузить решение (ZIP или Git URL)","operationId":"create_submission_api_v1_submissions_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_submission_api_v1_submissions_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/submissions/{submission_id}":{"get":{"tags":["submissions"],"summary":"Проверка по ID","operationId":"get_submission_api_v1_submissions__submission_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"submission_id","in":"path","required":true,"schema":{"type":"integer","title":"Submission Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/submissions/{submission_id}/status":{"get":{"tags":["submissions"],"summary":"Текущий статус проверки","operationId":"get_status_api_v1_submissions__submission_id__status_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"submission_id","in":"path","required":true,"schema":{"type":"integer","title":"Submission Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/submissions/{submission_id}/results":{"get":{"tags":["submissions"],"summary":"Детальные результаты чекеров","operationId":"get_results_api_v1_submissions__submission_id__results_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"submission_id","in":"path","required":true,"schema":{"type":"integer","title":"Submission Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/submissions/{submission_id}/rerun":{"post":{"tags":["submissions"],"summary":"Повторный запуск проверки","operationId":"rerun_submission_api_v1_submissions__submission_id__rerun_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"submission_id","in":"path","required":true,"schema":{"type":"integer","title":"Submission Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/submissions/{submission_id}/verdict":{"put":{"tags":["submissions"],"summary":"Вынести вердикт","operationId":"update_verdict_api_v1_submissions__submission_id__verdict_put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"submission_id","in":"path","required":true,"schema":{"type":"integer","title":"Submission Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerdictUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/submissions/{submission_id}/report":{"get":{"tags":["submissions"],"summary":"Скачать отчёт по проверке","operationId":"download_report_api_v1_submissions__submission_id__report_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"submission_id","in":"path","required":true,"schema":{"type":"integer","title":"Submission Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/submissions/{submission_id}/ai-review":{"get":{"tags":["submissions"],"summary":"AI-анализ решения","operationId":"ai_review_api_v1_submissions__submission_id__ai_review_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"submission_id","in":"path","required":true,"schema":{"type":"integer","title":"Submission Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/submissions/{submission_id}/events":{"get":{"tags":["submissions"],"summary":"SSE-поток статусов проверки","operationId":"submission_events_api_v1_submissions__submission_id__events_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"submission_id","in":"path","required":true,"schema":{"type":"integer","title":"Submission Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/candidates":{"get":{"tags":["candidates"],"summary":"Список кандидатов","operationId":"list_candidates_api_v1_candidates_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/candidates/{candidate_id}":{"get":{"tags":["candidates"],"summary":"Кандидат по ID","operationId":"get_candidate_api_v1_candidates__candidate_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"candidate_id","in":"path","required":true,"schema":{"type":"integer","title":"Candidate Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reports/stats":{"get":{"tags":["reports"],"summary":"Статистика проверок","operationId":"stats_api_v1_reports_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AssignmentCreate":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"checker_weights":{"additionalProperties":{"type":"number"},"type":"object","title":"Checker Weights"}},"type":"object","required":["title","description"],"title":"AssignmentCreate"},"AssignmentUpdate":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"checker_weights":{"anyOf":[{"additionalProperties":{"type":"number"},"type":"object"},{"type":"null"}],"title":"Checker Weights"}},"type":"object","title":"AssignmentUpdate"},"Body_create_submission_api_v1_submissions_post":{"properties":{"assignment_id":{"type":"integer","title":"Assignment Id"},"git_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Git Url"},"zip_file":{"anyOf":[{"type":"string","contentMediaType":"application/octet-stream"},{"type":"null"}],"title":"Zip File"}},"type":"object","required":["assignment_id"],"title":"Body_create_submission_api_v1_submissions_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginIn":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginIn"},"RegisterIn":{"properties":{"full_name":{"type":"string","maxLength":255,"minLength":2,"title":"Full Name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","minLength":6,"title":"Password"},"role":{"type":"string","enum":["candidate","expert","admin"],"title":"Role","default":"candidate"}},"type":"object","required":["full_name","email","password"],"title":"RegisterIn"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VerdictUpdate":{"properties":{"verdict":{"type":"string","enum":["approved","rejected","pending"],"title":"Verdict"}},"type":"object","required":["verdict"],"title":"VerdictUpdate"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}},"tags":[{"name":"auth","description":"Регистрация, логин, профиль и выход пользователя."},{"name":"assignments","description":"CRUD операций с тестовыми заданиями."},{"name":"submissions","description":"Загрузка решений, статусы проверки, результаты, вердикты, SSE и AI-review."},{"name":"candidates","description":"Работа со списком кандидатов и карточкой кандидата."},{"name":"reports","description":"Агрегированная статистика и рейтинг кандидатов."}]}